Set up OnKilled + cleanup
!!!!!
NOTE: The behavior of commands "gunner" and "driver" are not extensively
documented on the wiki. I'm not sure if respect/money increase will work
properly for kills from vehicles! I need to test these commands later
!!!!!
Cleaned up obsolete/already created functions from preInit

Created config values "DMS_BanditMoneyGainOnKill" and
"DMS_BanditRepGainOnKill" and "DMS_RemoveNVG"

OnKilled should now give respect and pop tabs to clients (if configured
to do so).

It should also give the onscreen popup for kills for respect and update
the values on the client (as well as updating the corresponding stats on
the server and database)

Use "MPKilled" EH instead of "Killed" for potential issues with
offloading.